Key value dictionary in Haskell.
This dictionary uses both a binary search tree and a trie like
structure to implement efficient key lookup for keys that share
prefixes between them, saving a lot of space. Keys are lists
of some type, and must have an order relation.

Dic is an abstract data type class with an interface to handle
dictionary operations and queries. The structure here implemented
supports that interface.